discovered major flaw in logic, changing direction
there is a major flaw in logic by doing sim[i] * R + sim[i+e] because
you get collisions and as such it incorrectly sorts. Thus need to go
back to previous methods where i do recursive search to get a first
working version of this.

Incidentally, even when this way works (when there are no collisions)
it is surprisingly slower then the msd radix sort--probably the
constant work is causing much slowdown (or another error).