fixed strings suffix nlogn method to be faster
It is much faster now, by taking shortcut and sorting buckets
as we go, rather than sorting all together, which makes for a
worse time because we resort all good ones all the time.
The buckets may be sorted faster with bucket sort, but as it is
we still have issues with the constant factors because we are
scanning from 0 to N multiple times.