adding 2 sat problem for the algo class
This version is solved by reducing the 2 sat problem
to finding the connected components in the graph build
by linking every pair of constraits in the graph:
(A v B) is equivalent to (~B -> A) and (~A -> B) and
we add both of later ones as links in the graph.
If there is any connected component that contains
A and ~A then the constraints are not satisfiable.