adding attempt a lockless circular queue
it does not seem to work in the atom processor to update
a long long int vi acompare and swap. it quite possibly could
work in 64bit, but if i want to get this to work here, i need
to be doing 32 bit swaps.
The reason i need 64 bit swap is to get the front and the count
to compare to avoid the ABA problem. However, if i use versioning
i may be able to get just that. I will try with 32bit versioning
next.