added above average problem, cryptography
above average is trivial and got it on first try.
Cryptography via java was not finishing on time,
I don't think there was a problem, but it was slow.
I had to use doubles in c so that it can be much faster.
Correctness is guaranteed up to a certain size of the
numbers, otherwise we get into not enough significant
digits to represent the numbers.
Interestingly, there is no need to round the final double.