removed unnecessary usage of set
put this back to c code, to compute the combinations
of the factors (a,b,c) with multiplicity (x,y,z)
all you need to do is multiply (x+1)*(y+1)*(z+1)
because you can take a 0,1,...x times (that's x+1)
and so on with b and c. When you take all for no times
that is no prime multiple, that leaves 1 as multiple.
Clever and no need for set to get the combinations.