Using the param as a lock does work but it's slow
Not sure if i can make this any faster, locks will probably
make this go faster, even though the idea is to make each
individual action fast by not calling into mutex lock.
Overall tho, because of the high concurrency that i have in
this case, it takes much longer because of the spinning activity
of many of the threads. This may perform better on a multiprocessor
machine because all the spinning will not prohibit the action thread
to do actual work.