Simplified logic to check for remaining producers
We already had counter atomically incremented, all needed to be
done is use that to identify how many threads had incremented it
and wait till that number reached NUMTHREAD above MAXSIZE.
So now the function is simpler to understand and works
in the same way without the need for two atomic icrements at the end.
Not that this is a concern for speed because at the end of the run
we sleep plenty to wait for the other threads to finish their work.